e617fe2 Fix various warnings (Luke Dashjr)

+++ src/rest.cpp
@@ -20,7 +20,7 @@
 
 using namespace std;
 
-static const int MAX_GETUTXOS_OUTPOINTS = 15; //allow a max of 15 outpoints to be queried at once
+static const size_t MAX_GETUTXOS_OUTPOINTS = 15; //allow a max of 15 outpoints to be queried at once
 
 enum RetFormat {
     RF_UNDEF,
